test_app.py
# --- Step 1: Add Imports ---
# (Keep all imports as they were in Stage 3)
import os
import google.generativeai as genai
from flask import Flask, render_template, request, redirect, url_for, flash, send_from_directory
from werkzeug.utils import secure_filename
from PIL import Image as PILImage
import io
import time
from datetime import datetime

# --- Step 2: Config Loading ---
# (Keep as is)
try:
    from config import (
        GEMINI_TEXT_API_KEY, GEMINI_IMAGE_API_KEY, SECRET_KEY, DATABASE_URI,
        UPLOAD_FOLDER, ALLOWED_EXTENSIONS, GEMINI_MODEL_NAME,
        DESCRIPTION_SIMILARITY_THRESHOLD, METADATA_SIMILARITY_THRESHOLD,
        IMAGE_SIMILARITY_THRESHOLD
    )
    print("Successfully imported from config.py")
except Exception as e: print(f"CRITICAL ERROR importing config: {e}"); exit()

# --- Step 3: Models Import and DB Initialization ---
# (Keep as is)
try:
    from models import db, Item
    print("Successfully imported from models.py")
except Exception as e: print(f"CRITICAL ERROR importing models: {e}"); exit()

# --- Flask App Setup ---
# (Keep as is)
app = Flask(__name__)
try:
    app.config['SECRET_KEY'] = SECRET_KEY
    app.config['SQLALCHEMY_DATABASE_URI'] = DATABASE_URI
    app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
    app.config['MAX_CONTENT_LENGTH'] = 16 * 1024 * 1024
    app.config['ALLOWED_EXTENSIONS'] = ALLOWED_EXTENSIONS
    print("Flask app configured.")
except Exception as e: print(f"CRITICAL ERROR setting Flask config: {e}"); exit()

# Initialize SQLAlchemy
# (Keep as is)
try:
    db.init_app(app)
    print("SQLAlchemy initialized with app.")
except Exception as e: print(f"CRITICAL ERROR: db.init_app(app) failed: {e}"); exit()

# --- Step 5: Gemini API Configuration ---
# (Keep as is)
print("--- Attempting Gemini API Configuration ---")
try:
    genai.configure(api_key=GEMINI_TEXT_API_KEY)
    print(f"Gemini API configured successfully (using text key ending ...{GEMINI_TEXT_API_KEY[-4:]}).")
except Exception as e: print(f"ERROR during genai.configure: {e}.")
text_model = None
vision_model = None
try:
    text_model = genai.GenerativeModel(GEMINI_MODEL_NAME)
    vision_model = genai.GenerativeModel(GEMINI_MODEL_NAME)
    print(f"Gemini models ('{GEMINI_MODEL_NAME}') initialized successfully.")
except Exception as e: print(f"ERROR creating Gemini models: {e}.")
print("--- Finished Gemini API Configuration Attempt ---")

# --- Step 4: Database Initialization Logic ---
# (Keep as is)
def init_db_test():
     with app.app_context():
        print("Attempting db.create_all()...")
        try: db.create_all(); print("db.create_all() executed.")
        except Exception as e: print(f"ERROR during db.create_all(): {e}")

# --- Step 6: Context Processor ---
# (Keep as is)
print("Adding context processor...")
@app.context_processor
def inject_now():
    from datetime import datetime
    return {'now': datetime.utcnow}
print("Context processor added.")

# --- Routes ---

@app.route('/')
def index():
    """Renders the real home page."""
    print("Request received for '/' route. Attempting to render index.html...")
    try:
        return render_template('index.html')
    except Exception as e:
        print(f"!!! ERROR occurred during render_template('index.html'): {e}")
        raise e

# --- Step 8: ADD PLACEHOLDER ROUTES ---
@app.route('/report_found') # Define the route URL
def report_found():         # Define the function with the endpoint name
    print("Placeholder route '/report_found' called.")
    # For now, just return simple text. Later, add the real logic.
    return "Placeholder page for Reporting Found Items"

@app.route('/search_lost') # Define the route URL
def search_lost():          # Define the function with the endpoint name
    print("Placeholder route '/search_lost' called.")
    return "Placeholder page for Searching Lost Items"

# Add placeholder for uploads route as well if base.html or others might link images soon
@app.route('/uploads/<path:filename>')
def uploaded_file(filename):
     print(f"Placeholder route '/uploads/{filename}' called.")
     # In a real app, this would serve the file. For now, return 404 or text.
     # return send_from_directory(app.config['UPLOAD_FOLDER'], filename)
     return f"Would serve file: {filename}", 404 # Return 404 to avoid actual file serving yet

# -----------------------------------------

# --- Run Application ---
if __name__ == '__main__':
    print("--- Starting Test App Stage 4 ---") # Stage 4 now

    if not os.path.exists(UPLOAD_FOLDER):
        try: os.makedirs(UPLOAD_FOLDER); print(f"Created upload directory: {UPLOAD_FOLDER}")
        except OSError as e: print(f"Warning: Error creating upload directory {UPLOAD_FOLDER}: {e}")

    init_db_test()

    print("Starting Flask development server for test...")
    # Use a different port again
    app.run(debug=True, host='0.0.0.0', port=5005) # Port 5005
    print("--- Test App Stage 4 Stopped ---")