requirements.txt
Flask
Flask-SQLAlchemy
google-generativeai
Pillow
python-dotenv  # Good practice, though keys are in config.py for this example
werkzeug       # For secure filenames